# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Gatka Association India API
  version: 1.0.0
  description: |
    REST API for centralized Gatka management (roles, geography, registrations, Razorpay, competitions, camps, attendance, results, notices, migration).

    **Banners:** carousel rows scoped by **state**. Public carousel **requires a state filter**: `GET /banners/public?stateId=…` or `GET /banners/public/by-state/{stateId}`. Rows have `imageUrl`, optional `title` / `subtitle`, `sortOrder`. National/state admins use `/banners/admin` (`stateId` query optional for nationals).

    **Uploads:** Cloudflare R2 — `POST /uploads/r2/upload` (multipart) and `POST /uploads/r2/presign-put` (JSON) are **public** (no JWT); protected by **rate limits**. Store returned `publicUrl` on entities (e.g. banners, registration photos).

    **Rate limiting:** per-IP limits on `/api` (and stricter limits on auth and upload presign). Exceeded limits return **429** with JSON `{ error, code: RATE_LIMITED, kind, retryAfterSec }`. Use `TRUST_PROXY=1` behind a reverse proxy.

    **Age categories:** `GET /age-categories` is **unauthenticated** (no JWT).

    **Attendance:** `POST /attendance` requires COACH, TRAINING_CENTER, or district/state/national admin.
    `GET /attendance/user/{userId}` is for the same user when `role` is PLAYER, or any user for other roles.
    `GET /attendance/competition/{competitionId}/summary` lists TOURNAMENT rows for that competition (district/state/national admin only).
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    StateDomain:
      type: object
      required: [stateId, stateName, domainName]
      properties:
        stateId: { type: string }
        stateName: { type: string }
        domainName: { type: string }
    StateDomainPatch:
      type: object
      required: [domainName]
      properties:
        domainName:
          type: string
          description: Domain hostname only (example `pb.dummy.gatka.in`), no protocol/path.
...
paths:
...
  /domains:
    get:
      security: [{ bearerAuth: [] }]
      summary: List state domains (national admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: boolean }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/StateDomain" }
        "403":
          description: Forbidden (NATIONAL_ADMIN only)
  /domains/{stateId}:
    patch:
      security: [{ bearerAuth: [] }]
      summary: Update state domain by state id (national admin)
      parameters:
        - in: path
          name: stateId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StateDomainPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: boolean }
                  message: { type: string }
                  data: { $ref: "#/components/schemas/StateDomain" }
        "404":
          description: State not found
        "409":
          description: Domain already assigned to another state
...